Use `size_t` for counter

+++ esphome/components/nextion/nextion.cpp
@@ -101,8 +101,8 @@ bool Nextion::check_connect_() {
   ESP_LOGN(TAG, "connect: %s", response.c_str());
 
   // Parse the comok response fields directly, avoiding a heap-allocated vector.
-  uint8_t field_count = 0;
-  uint8_t field_idx = 0;
+  size_t field_count = 0;
+  size_t field_idx = 0;
   size_t start = 0;
   size_t end = 0;
   while ((start = response.find_first_not_of(',', end)) != std::string::npos) {
